# src/server/src/resources/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the API contract for the server for the DMS LEAP application",
    "version": "1.0.0",
    "title": "LEAP DMS APIs",
    "termsOfService": ""
  },
...
  "paths": {
...
    "/users/byEmailId/{emailId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by email Id",
        "description": "Get details of user based on its email ID",
        "operationId": "getUserByEmailId",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emailId",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ByEmailId"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/messages": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Create thread by researcher",
        "description": "Create new message thread based on data passed in the request body",
        "operationId": "createThread",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Messages"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get thread with unread message count",
        "description": "Get thread with unread message count as per query parameter",
        "operationId": "getThreads",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "AppId"
            }
          },
          {
            "in": "query",
            "name": "appId",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Page"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "schema": {
              "type": "integer",
              "description": "Size"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "description": "Size"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesId"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
    "/messages/{id}/participants": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Fetch participants information",
        "description": "Fetch participants information based on threadId",
        "operationId": "getParticipants",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of thread to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "userId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Participant"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Participant not found"
          }
        }
      }
    },
    "/messages/{id}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Fetch chat between sender and recipient",
        "description": "Fetch chats for the perticular threads betweeen sender Id (from headers) and recipient (Query params)",
        "operationId": "getChats",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of thread to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "name": "senderId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "senderId"
            }
          },
          {
            "in": "query",
            "name": "recipientId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "RecepientId"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "schema": {
              "type": "integer",
              "description": "Size"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "description": "Size"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatId"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      },
      "put": {
        "tags": [
          "Messages"
        ],
        "summary": "Add chat message",
        "description": "Add chat message in existing thread",
        "operationId": "createChat",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of thread to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "schema": {
              "type": "integer",
              "description": "Size"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "description": "Size"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chats"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      }
    },
    "/messages/{id}/markAsRead": {
      "put": {
        "tags": [
          "Messages"
        ],
        "summary": "Mark read chat from threads",
        "description": "Mark chat as  from threads based on Id",
        "operationId": "markAsRead",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of thread to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "recipientId",
            "name": "recipientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "description": "senderId",
            "name": "senderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMarkAsRead"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
    "/messages/unreadCount": {
      "put": {
        "tags": [
          "Messages"
        ],
        "summary": "Get unread count of message",
        "description": "Get unread count of message",
        "operationId": "unreadCount",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "userId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/unreadCountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
    "/messages/titleSearch": {
      "put": {
        "tags": [
          "Messages"
        ],
        "summary": "Get unread count of message",
        "description": "Get unread count of message",
        "operationId": "titleSearch",
        "parameters": [
          {
            "in": "query",
            "description": "appId",
            "name": "appId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "message search string",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "userId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "description": "userId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "userId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesId"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://leapapi-dev.leap.dellmed.utexas.edu/api"
    },
    {
      "url": "https://localhost:8080/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "description": "message details",
        "properties": {
          "app": {
            "type": "string",
            "example": "5fec6f918048d25b1c189a2e"
          },
          "title": {
            "type": "string",
            "example": "test study"
          },
          "userName": {
            "type": "string",
            "example": "firstname lastName"
          },
          "userId": {
            "type": "string",
            "example": "5ff40edee62fe27178eab167"
          }
        }
      },
      "Messages": {
        "type": "object",
        "description": "message details",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60c33dd35fcf392a0165748d"
          },
          "app": {
            "type": "string",
            "example": "5fec6f6e8048d25b1c189a2d"
          },
          "title": {
            "type": "string",
            "example": "test study"
          },
          "createdBy": {
            "type": "string",
            "example": "firstname lastName"
          },
          "updatedBy": {
            "type": "string",
            "example": "firstname lastname"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-11T08:31:42.270Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-11T08:31:42.270Z"
          },
          "chats": {
            "type": "array",
            "default": [],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Participant": {
        "type": "array",
        "items": {
          "type": "object",
          "description": "participant details",
          "properties": {
            "id": {
              "type": "string"
            },
            "user": {
              "type": "string"
            },
            "count": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "example": [
          {
            "id": "5ff559f333aad739800b9bc4",
            "user": "firstname lastname",
            "count": 1
          },
          {
            "id": "5ff5416533aad739800b9bbf",
            "user": "firstname lastname",
            "count": 2
          }
        ]
      },
      "Chat": {
        "type": "object",
        "description": "chat details",
        "properties": {
          "sender": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "5ff40edee62fe27178eab165"
              },
              "name": {
                "type": "string",
                "example": "firstname lastname"
              }
            }
          },
          "recipient": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "5ff5416533aad739800b9bbf"
              },
              "name": {
                "type": "string",
                "example": "firstname lastname"
              }
            }
          },
          "body": {
            "type": "string",
            "example": "Hi, Good Morning"
          }
        }
      },
...
      "ChatMarkAsRead": {
        "type": "object",
        "description": "chat details",
        "properties": {
          "message": {
            "type": "string",
            "example": "chats mark as a read successfully"
          }
        }
      },
...
      "ByEmailId": {
        "type": "object",
        "properties": {
          "isBiometricEnabled": {
            "type": "boolean",
            "example": false
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "_id": {
            "type": "string",
            "example": "607d7b1e2cefd922906ae35e"
          },
          "keycloakId": {
            "type": "string",
            "example": "9d2efd8d-7591-4cfb-ac6a-36da63c904bb"
          },
          "email": {
            "type": "string",
            "example": "balwant_padwal@persistent.com"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "inviteCode": {
            "type": "string",
            "example": "balwant"
          },
          "createdBy": {
            "type": "string",
            "example": "Admin User"
          },
          "updatedBy": {
            "type": "string",
            "example": "Balwant Padwal"
          },
          "createdAt": {
            "type": "string",
            "example": "2021-04-19T12:44:14.893Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2021-04-19T13:06:39.150Z"
          },
          "apps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "isActive": {
                  "type": "boolean"
                },
                "consentedMedicalRecords": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "5ffc32eda3c2ec783e43b628"
                  }
                },
                "_id": {
                  "type": "string",
                  "example": "5ff48ef6a4bcf442483d6f48"
                },
                "app": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "5fec6f6e8048d25b1c189a2d"
                    },
                    "isDeleted": {
                      "type": "boolean",
                      "example": false
                    },
                    "isActive": {
                      "type": "boolean",
                      "example": true
                    },
                    "notifications": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "5ff40a5a061e3f78d0f36e42,6006868fcb4b3041d0157393"
                      }
                    },
                    "appName": {
                      "type": "string",
                      "example": "studyApp"
                    },
                    "longDescription": {
                      "type": "string",
                      "example": "long description of app"
                    },
                    "createdBy": {
                      "type": "string",
                      "example": "Balwant"
                    },
                    "updatedBy": {
                      "type": "string",
                      "example": "Balwant"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2020-12-30T12:10:26.531Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2020-12-30T12:10:26.531Z"
                    },
                    "shortDescription": {
                      "type": "string",
                      "example": "short description of app"
                    },
                    "appLogo": {
                      "type": "string",
                      "example": "studyapp_logo.svg"
                    },
                    "appUrl": {
                      "type": "string",
                      "example": ""
                    },
                    "consentVersion": {
                      "type": "string",
                      "example": "v4"
                    },
                    "consentTermOfUse": {
                      "type": "string",
                      "example": "test contents for consent terms of use."
                    },
                    "consentPrivacy": {
                      "type": "string",
                      "example": "test contents for consent privacy."
                    },
                    "consentPolicy": {
                      "type": "string",
                      "example": "<div>Consent policy form</div>"
                    },
                    "medicalRecords": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "5ffc32eda3c2ec783e43b628"
                      }
                    }
                  }
                },
                "consentUpdatedAt": {
                  "type": "string",
                  "example": "2021-04-19T15:19:45.679Z"
                }
              }
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "isPushed": {
                  "type": "boolean"
                },
                "isRead": {
                  "type": "boolean"
                },
                "_id": {
                  "type": "string"
                },
                "notification": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                }
              }
            },
            "example": [
              {
                "isPushed": false,
                "isRead": false,
                "_id": "5ff4a85eb4c7a14cec131a7c",
                "notification": "5ff40a5a061e3f78d0f36e42",
                "createdAt": "2021-01-05T17:57:06.586Z"
              },
              {
                "isPushed": false,
                "isRead": false,
                "_id": "5ff4a85eb4c7a14cec131a7c",
                "notification": "6006868fcb4b3041d0157393",
                "createdAt": "2021-01-05T17:57:06.586Z"
              }
            ]
          },
          "lastLoginTime": {
            "type": "string",
            "example": "2021-04-19T15:56:59.647Z"
          },
          "birthday": {
            "type": "string",
            "example": "2021-04-19T00:00:00.000Z"
          },
          "firstName": {
            "type": "string",
            "example": "Balwant"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "genderOther": {
            "type": "string",
            "example": null
          },
          "lastName": {
            "type": "string",
            "example": "Padwal"
          },
          "phoneNumberPrimary": {
            "type": "string",
            "example": 1111111111
          },
          "phoneNumberSecondary": {
            "type": "string",
            "example": 1111111111
          },
          "zip": {
            "type": "string",
            "example": 12345
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}